requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymupdf==1.23.8
rank-bm25==0.2.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
numpy==1.24.3
python-dotenv==1.0.0

